+++ lib/haml_lint.rb
@@ -16,6 +16,12 @@ require 'haml_lint/version'
 
 require 'haml'
 
+# Load all parse tree node classes
+require 'haml_lint/tree/node'
+Dir[File.expand_path('haml_lint/tree/*.rb', File.dirname(__FILE__))].each do |file|
+  require file
+end
+
 # Load all linters
 Dir[File.expand_path('haml_lint/linter/*.rb', File.dirname(__FILE__))].each do |file|
   require file