+++ lib/haml_lint/linter/leading_comment_space.rb
@@ -4,7 +4,7 @@ module HamlLint
     include LinterRegistry
 
     def visit_haml_comment(node)
-      return if node.value[:text][0] == ' '
+      return if node.text[0] == ' '
 
       add_lint(node, 'Comment should have a space after the `#`')
     end