+++ lib/haml_lint/tree/comment.rb
@@ -1,0 +1,5 @@
+module HamlLint
+  # Represents a visible XHTML comment in a HAML document.
+  class CommentNode
+  end
+end