+++ lib/haml_lint/tree/doctype_node.rb
@@ -1,0 +1,5 @@
+module HamlLint
+  # Represents a doctype definition for a HAML document.
+  class DoctypeNode
+  end
+end