+++ lib/haml_lint/tree/haml_comment_node.rb
@@ -1,0 +1,5 @@
+module HamlLint::Tree
+  # Represents a HAML comment node.
+  class HamlCommentNode < Node
+  end
+end